# docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  github_com_Ulpio_vergo_internal_domain_project.Project:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      org_id:
        type: string
      updated_by:
        type: string
    type: object
...
  internal_http_handlers.ErrorDetailResponse:
    description: Error response with detail
    properties:
      detail:
        example: duplicate key
        type: string
      error:
        example: create_fail
        type: string
    type: object
  internal_http_handlers.ErrorResponse:
    description: Standard error response
    properties:
      error:
        example: not_found
        type: string
    type: object
...
  internal_http_handlers.ProjectIn:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Multi-tenant SaaS boilerplate API with JWT auth, RBAC, and S3 storage.
  title: Vergo API
  version: 0.1.0
paths:
...
  /projects:
    get:
      parameters:
      - description: Organization ID
        in: header
        name: X-Org-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: items
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/github_com_Ulpio_vergo_internal_domain_project.Project'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorDetailResponse'
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: header
        name: X-Org-ID
        required: true
        type: string
      - description: Project data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.ProjectIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Ulpio_vergo_internal_domain_project.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorDetailResponse'
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - Projects
  /projects/{id}:
    delete:
      parameters:
      - description: Organization ID
        in: header
        name: X-Org-ID
        required: true
        type: string
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - Projects
    get:
      parameters:
      - description: Organization ID
        in: header
        name: X-Org-ID
        required: true
        type: string
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Ulpio_vergo_internal_domain_project.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project
      tags:
      - Projects
    patch:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: header
        name: X-Org-ID
        required: true
        type: string
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated project data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.ProjectIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Ulpio_vergo_internal_domain_project.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorDetailResponse'
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - Projects
...
securityDefinitions:
  BearerAuth:
    description: Bearer token (e.g. "Bearer eyJhbG...")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"